Break apart AdapterSpec, Prompt, RequestState and ScenarioState from adapter.py
