+++ src/helm/benchmark/adaptation/adapter.py
@@ -1,16 +1,19 @@
 import random
-from dataclasses import dataclass, field, replace
+from dataclasses import dataclass, replace
 from itertools import cycle
 from typing import List, Dict, Tuple, Optional, Callable
-from collections import defaultdict, OrderedDict
-import re
+from collections import defaultdict
 
 import numpy as np
 
-from helm.common.general import serialize, indent_lines, format_text_lines, parallel_map, flatten_list
+from helm.common.general import parallel_map, flatten_list
 from helm.common.hierarchical_logger import hlog, htrack, htrack_block
-from helm.common.request import Request, RequestResult
+from helm.common.request import Request
 from helm.common.tokenization_request import TokenizationToken
+from helm.benchmark.adaptation.adapter_spec import AdapterSpec
+from helm.benchmark.adaptation.prompt import Prompt
+from helm.benchmark.adaptation.request_state import RequestState
+from helm.benchmark.adaptation.scenario_state import ScenarioState
 from helm.benchmark.scenarios.scenario import Instance, TRAIN_SPLIT, EVAL_SPLITS, CORRECT_TAG, Input
 from helm.benchmark.window_services.window_service import WindowService, EncodeResult
 from helm.benchmark.window_services.window_service_factory import WindowServiceFactory
@@ -30,251 +33,6 @@ ADAPT_MULTIPLE_CHOICE_METHODS = [ADAPT_MULTIPLE_CHOICE_SEPARATE_ORIGINAL, ADAPT_
 # TODO: It would be better if we read the following from the adapter spec.
 RANKING_CORRECT_LABEL = "Yes"
 RANKING_WRONG_LABEL = "No"
-
-
-@dataclass(frozen=True)
-class Substitution:
-    """Represents a regular expression search/replace."""
-
-    source: str
-    target: str
-
-
-# TODO: move into separate file
-@dataclass(frozen=True)
-class AdapterSpec:
-    """
-    Specifies how to take a `Scenario` (a list of `Instance`s) and produce a
-    `ScenarioState` (set of `Request`s ). Instead of having free-form prompt
-    hacking, we try to make the process more declarative and systematic.
-    Note that an `Instance` could produce many `Request`s (e.g., one for each `Reference`).
-    """
-
-    # Method of adaptation
-    method: str = ""
-
-    # Prepend all prompts with this string.
-    # For example, it is recommended to prefix all prompts with [NLG] for UL2.
-    global_prefix: str = ""
-
-    # Prompt starts with instructions
-    instructions: str = ""
-
-    # What goes before the input
-    input_prefix: str = "Input: "
-
-    # What goes after the input
-    input_suffix: str = "\n"
-
-    # What goes before the input (for multiple choice)
-    reference_prefix: str = "A. "
-
-    # What goes before the input (for multiple choice)
-    reference_suffix: str = "\n"
-
-    # What goes before the output
-    output_prefix: str = "Output: "
-
-    # What goes after the output
-    output_suffix: str = "\n"
-
-    # What goes between instruction and in-context example blocks in the constructed prompt
-    instance_prefix: str = "\n"
-
-    # List of regular expression substitutions that we perform
-    substitutions: List[Substitution] = field(default_factory=list, hash=False)
-
-    # Maximum number of (in-context) training instances to put into the prompt
-    max_train_instances: int = 5
-
-    # Maximum number of evaluation instances. For getting valid numbers, this
-    # should be the entire dataset; only reduce this for piloting.
-    max_eval_instances: Optional[int] = None
-
-    # Generate this many outputs (which could be realized by `num_completions`
-    # or `top_k_per_token`).
-    num_outputs: int = 5
-
-    # Number of trials, where in each trial we choose an independent, random
-    # set of training instances.  Used to compute error bars.
-    num_train_trials: int = 1
-
-    # Decoding parameters (inherited by `Request`)
-
-    # Model to make the request to (need to fill in)
-    model: str = ""
-
-    # Temperature to use
-    temperature: float = 1
-
-    # Maximum number of tokens to generate
-    max_tokens: int = 100
-
-    # When to stop (set hash=False to make `AdapterSpec` hashable)
-    stop_sequences: List[str] = field(default_factory=list, hash=False)
-
-    # Random string (used concretely to bypass cache / see diverse results)
-    random: Optional[str] = None
-
-
-# TODO: move into separate file
-@dataclass(frozen=True)
-class RequestState:
-    """
-    A `RequestState` represents a single `Request` made on behalf of an `Instance`.
-    It should have all the information that's needed later for a `Metric` to be
-    able to understand the `Request` and its `RequestResult`.
-    """
-
-    instance: Instance
-    """Which instance we're evaluating"""
-
-    reference_index: Optional[int]
-    """Which reference of the instance we're evaluating (if any)"""
-
-    request_mode: Optional[str]
-    """Which request mode ("original" or "calibration") of the instance we're evaluating (if any)
-    (for ADAPT_MULTIPLE_CHOICE_SEPARATE_CALIBRATED)"""
-
-    train_trial_index: int
-    """Which training set this request is for"""
-
-    output_mapping: Optional[Dict[str, str]]
-    """How to map the completion text back to a real output (e.g., for multiple choice, "B" => "the second choice")"""
-
-    request: Request
-    """The request that is actually made"""
-
-    result: Optional[RequestResult]
-    """The result of the request (filled in when the request is executed)"""
-
-    num_train_instances: int
-    """Number of training instances (i.e., in-context examples)"""
-
-    prompt_truncated: bool
-    """Whether the prompt (instructions + test input) is truncated to fit the model's context window."""
-
-    num_conditioning_tokens: int = 0
-    """The number of initial tokens that will be ignored when computing language modeling metrics"""
-
-    def render_lines(self) -> List[str]:
-        output = [f"train_trial_index: {self.train_trial_index}"]
-        if self.reference_index:
-            output.append(f"reference_index: {self.reference_index}")
-
-        output.append("instance {")
-        output.extend(indent_lines(self.instance.render_lines()))
-        output.append("}")
-
-        # Part of request but render multiline
-        output.append("request.prompt {")
-        output.extend(indent_lines(format_text_lines(self.request.prompt)))
-        output.append("}")
-
-        output.append("request {")
-        output.extend(indent_lines(serialize(self.request)))
-        output.append("}")
-
-        if self.result:
-            output.append("result {")
-            output.extend(indent_lines(self.result.render_lines()))
-            output.append("}")
-
-        return output
-
-
-# TODO: move into separate file
-@dataclass
-class ScenarioState:
-    """
-    A `ScenarioState` represents the output of adaptation.  Contains a set of
-    `RequestState` that were created and executed (a `ScenarioState` could be
-    pre-execution or post-execution).
-    """
-
-    # What strategy we used for adaptation
-    adapter_spec: AdapterSpec
-
-    # List of `RequestState`s that were produced by adaptation (and execution)
-    request_states: List[RequestState]
-
-    def __post_init__(self):
-        # Create derived indices based on `request_states` so it's easier for
-        # the `Metric` later to access them.  Two things are produced:
-        self.request_state_map: Dict[Tuple[int, Instance, Optional[int]], List[RequestState]] = defaultdict(list)
-
-        # Python doesn't support an ordered set, so use an OrderedDict instead to maintain insertion order
-        instances_set: Dict[Instance, None] = OrderedDict()
-        for request_state in self.request_states:
-            instances_set[request_state.instance] = None
-            key = (request_state.train_trial_index, request_state.instance, request_state.reference_index)
-            self.request_state_map[key].append(request_state)
-        self.instances: List[Instance] = list(instances_set.keys())
-
-    def get_request_states(
-        self, train_trial_index: int, instance: Instance, reference_index: Optional[int]
-    ) -> List[RequestState]:
-        return self.request_state_map.get((train_trial_index, instance, reference_index), [])
-
-
-@dataclass(frozen=True)
-class Prompt:
-    """Result of prompt construction."""
-
-    # Global prefix, carried over from `AdapterSpec`
-    global_prefix: str
-
-    # Instance prefix, carried over from `AdapterSpec`
-    instance_prefix: str
-
-    # Substitutions, carried over from `AdapterSpec`
-    substitutions: List[Substitution]
-
-    # Instructions for the task
-    instructions_block: str
-
-    # Train instance blocks for the prompt
-    train_instance_blocks: List[str]
-
-    # Evaluation instance
-    eval_instance_block: str
-
-    # If the prompt (instructions + test input) needs to be truncated to fit the model's context window,
-    # this is the truncated text.
-    truncated_text: Optional[str] = None
-
-    @property
-    def text(self) -> str:
-        # Text for the prompt, might be truncated
-        if self.truncated_text:
-            return self.truncated_text
-
-        # Construct non-truncated input
-        blocks: List[str] = (
-            ([self.instructions_block] if self.instructions_block else [])
-            + self.train_instance_blocks
-            + [self.eval_instance_block]
-        )
-        non_truncated_text: str = self.instance_prefix.join(blocks)
-
-        # Note: this could be implemented via substitutions.
-        if self.global_prefix:
-            non_truncated_text = f"{self.global_prefix} {non_truncated_text}"
-
-        # Perform substitutions (e.g., add "<br>" before "\n")
-        for subst in self.substitutions:
-            non_truncated_text = re.sub(subst.source, subst.target, non_truncated_text)
-
-        return non_truncated_text
-
-    @property
-    def truncated(self) -> bool:
-        return self.truncated_text is not None
-
-    @property
-    def num_train_instances(self) -> int:
-        # Number of training instances in the prompt
-        return len(self.train_instance_blocks)
 
 
 @dataclass(frozen=True)
