+++ src/helm/benchmark/metrics/metric.py
@@ -10,12 +10,10 @@ from helm.benchmark.augmentations.perturbation_description import (
     PERTURBATION_ORIGINAL,
     PERTURBATION_WORST,
 )
-from helm.benchmark.adaptation.adapter import (
-    AdapterSpec,
-    ScenarioState,
-    RequestState,
-    ADAPT_LANGUAGE_MODELING,
-)
+from helm.benchmark.adaptation.adapter import ADAPT_LANGUAGE_MODELING
+from helm.benchmark.adaptation.scenario_state import ScenarioState
+from helm.benchmark.adaptation.request_state import RequestState
+from helm.benchmark.adaptation.adapter_spec import AdapterSpec
 from helm.benchmark.scenarios.scenario import Instance
 from .metric_name import MetricName, MetricContext
 from .metric_service import MetricService
