+++ src/helm/benchmark/scenarios/numeracy_scenario.py
@@ -12,7 +12,8 @@ from sympy import Symbol, Poly, diff
 from sympy.parsing.sympy_parser import standard_transformations, implicit_multiplication_application
 from typing import List, Optional, Tuple, Dict
 
-from helm.benchmark.adaptation.adapter import AdapterSpec, ADAPT_GENERATION
+from helm.benchmark.adaptation.adapter import ADAPT_GENERATION
+from helm.benchmark.adaptation.adapter_spec import AdapterSpec
 from helm.benchmark.window_services.tokenizer_service import TokenizerService
 from helm.common.authentication import Authentication
 from helm.proxy.services.server_service import ServerService
